+++ optional-tooling/python/novel_distiller/exporters/__init__.py
@@ -1,8 +1,0 @@
-"""
-Exporters package
-"""
-
-from .json_exporter import JsonExporter
-from .markdown_exporter import MarkdownExporter
-
-__all__ = ["JsonExporter", "MarkdownExporter"]