+++ optional-tooling/python/novel_distiller/loaders/__init__.py
@@ -1,9 +1,0 @@
-"""
-Loaders package
-"""
-
-from .txt_loader import TxtLoader
-from .epub_loader import EpubLoader
-from .chapter_splitter import ChapterSplitter
-
-__all__ = ["TxtLoader", "EpubLoader", "ChapterSplitter"]