+++ optional-tooling/python/novel_distiller/utils/__init__.py
@@ -1,7 +1,0 @@
-"""
-Utils package
-"""
-
-from .llm_client import LLMClient
-
-__all__ = ["LLMClient"]