+++ optional-tooling/python/novel_distiller/visualizers/__init__.py
@@ -1,7 +1,0 @@
-"""
-Visualizers package
-"""
-
-from .relationship_visualizer import RelationshipVisualizer
-
-__all__ = ["RelationshipVisualizer"]